pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"
autonity = "^4.0.0"
asgiref = "^3.8.1"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"
django-crontab = "^0.7.1"
psycopg2 = "^2.9.3"
django-filter = "^22.1"
pydantic = "^2.7.3"
pynacl = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
